Re-wrote the throttle manager to better support multiple concurrent streams.
New design is based on using byte tokens, so it active from the start, instead of using a short measure window.

The measure window was problematic on really fast streams as the initial writes could exceed the limit many times, causing extreme waits to be issued.

With the token based approach, a stream can use more than allocated, but this causes delays from the begining so the wait periods do not swing wildly.

The new throttle manager does not work super if the requested chunks are much larger than the limit (say limit 10kb/s and write a 100mb chunk, will cause a very large delay).

Default in .NET seems to be 64kb, which gives reasonable delays.